Avoid breaking changes on current contract creation model

+++ tests/op_datasets/etl/intermediate/models/test_contract_creation_traces.py
@@ -87,7 +87,7 @@ class TestCreationTraces001(IntermediateModelTestBase):
             "error": "VARCHAR",
             "status": "BIGINT",
             "tx_method_id": "VARCHAR",
-            "code_bytelength": "INTEGER",
+            "code_bytelength": "DOUBLE",
         }
 
     def test_single_tx(self):
